Move getFilteredFreaks function to separate module
