+++ src/components/FreaksGrid/FreaksGrid.jsx
@@ -3,6 +3,7 @@ import { useSelector } from "react-redux";
 import PropTypes from 'prop-types';
 import FreakTile from '../FreakTile/FreakTile';
 import './FreaksGrid.scss';
+import { getFilteredFreaks } from '../../filters/freaksFilter'
 import { PhotoModelProps } from '../../models/freak';
 
 function FreaksGrid({ freaks }) {
@@ -18,47 +19,7 @@ function FreaksGrid({ freaks }) {
         />
       ))};
 
-      const getFilteredFreaks = () => {
-        const getSkillFilteredFreaks = () => {
-          if (filters.skills.length === 0) {
-            return freaks;
-          }
-
-          const isFilteredTech = tech => {
-            return filters.skills.some(filter => filter === tech.name)
-          };
-
-          const hasFilteredTech = freak => {
-            return freak.technologies.some(tech => isFilteredTech(tech))
-          };
-
-          return freaks.filter( freak => hasFilteredTech(freak));
-        };
-
-        const getProjectFilteredFreaks = () => {
-          if (filters.projects.length === 0) {
-            return freaks;
-          }
-
-          const isFilteredProj = proj => {
-            return filters.projects.some(filter => filter === proj.name)
-          };
-
-          const hasFilteredProj = freak => {
-            return freak.projects.some(tech => isFilteredProj(tech))
-          };
-
-          return freaks.filter( freak => hasFilteredProj(freak));
-        };
-
-        const intersection = (arr1, arr2) => {
-          return arr1.filter(x => arr2.includes(x));
-        };
-
-        return intersection(getSkillFilteredFreaks(), getProjectFilteredFreaks());
-      }
-
-      return getFreakTiles(getFilteredFreaks(freaks));
+      return getFreakTiles(getFilteredFreaks(freaks, filters));
   }
 
   const [tiles, setTiles] = useState(getTiles);
